refactor: consolidate base64 and file-content helpers in core
