+++ .changeset/small-files-share.md
@@ -1,0 +1,8 @@
+---
+'generaltranslation': minor
+'@generaltranslation/api': minor
+'gt': patch
+'gt-sanity': patch
+---
+
+Consolidate base64 and file-content helpers in `generaltranslation/internal`, remove the generated API package's workspace dependency surface, and reuse the core helpers in the CLI and Sanity integration.
