+++ packages/api/src/index.ts
@@ -5,12 +5,6 @@ export type {
   GetJobStatuses,
   JobResult,
 } from './wrappers/awaitJobs';
-export {
-  decodeBase64,
-  decodeFileContent,
-  encodeBase64,
-  encodeFileContent,
-} from './wrappers/base64';
 export { DEFAULT_BATCH_SIZE, processBatches } from './wrappers/batch';
 export type { BatchOptions } from './wrappers/batch';
 export { API_VERSION, createApiClient } from './wrappers/client';
