+++ packages/sanity/src/serialization/data.ts
@@ -1,3 +1,5 @@
+import { decode, encode } from 'generaltranslation/internal';
+
 export function attachGTData(
   html: string,
   data: Record<string, unknown>,
@@ -57,26 +59,4 @@ export function detachGTData(html: string): {
     html: firstElement.outerHTML,
     data: extractedData,
   };
-}
-
-// Encode a string to base64
-export function encode(data: string): string {
-  // Browser path
-  const bytes = new TextEncoder().encode(data);
-  let binary = '';
-  for (let i = 0; i < bytes.length; i++) {
-    binary += String.fromCharCode(bytes[i]);
-  }
-  return btoa(binary);
-}
-
-// Decode a base64 string to a string
-export function decode(base64: string): string {
-  // Browser path
-  const binary = atob(base64);
-  const bytes = new Uint8Array(binary.length);
-  for (let i = 0; i < binary.length; i++) {
-    bytes[i] = binary.charCodeAt(i);
-  }
-  return new TextDecoder().decode(bytes);
 }
